api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "python-dotenv",
  "httpx",
  "PyJWT",
]
